Split hit detection
Split expensive hit detection into its own class.
