+++ tests/utilities.py
@@ -4,6 +4,7 @@ import gymnasium as gym
 import retro
 from triforce.action_space import ZeldaActionSpace
 from triforce.objective_selector import ObjectiveSelector
+from triforce.zelda_hit_detect import ZeldaHitDetect
 from triforce.zelda_wrapper import ZeldaGameWrapper
 
 class CriticWrapper(gym.Wrapper):
@@ -50,6 +51,7 @@ class ZeldaActionReplay:
         env = retro.make(game='Zelda-NES', state=savestate, inttype=retro.data.Integrations.CUSTOM_ONLY, render_mode=render_mode)
         self.data = env.data
         env = ZeldaGameWrapper(env, deterministic=True)
+        env = ZeldaHitDetect(env)
         env = ObjectiveSelector(env)
         env = ZeldaActionSpace(env, 'all')
         if wrapper:
