+++ triforce/zelda_wrapper.py
@@ -13,8 +13,8 @@ import gymnasium as gym
 import numpy as np
 
 from .zelda_game_data import zelda_game_data
-from .zelda_game import AnimationState, Direction, TileState, ZeldaEnemy, get_bomb_state, is_health_full, is_in_cave, \
-                        is_link_stunned, is_mode_death, get_beam_state, is_mode_scrolling, ZeldaObjectData, \
+from .zelda_game import AnimationState, Direction, TileState, ZeldaEnemy, is_health_full, is_in_cave, \
+                        is_link_stunned, get_beam_state, is_mode_scrolling, ZeldaObjectData, \
                         is_room_loaded, is_sword_frozen, get_heart_halves, position_to_tile_index, tiles_to_weights
 from .model_parameters import MAX_MOVEMENT_FRAMES, ATTACK_COOLDOWN, ITEM_COOLDOWN, CAVE_COOLDOWN, WS_ADJUSTMENT_FRAMES
 
@@ -54,8 +54,6 @@ class ZeldaGameWrapper(gym.Wrapper):
         self._location = None
         self._last_info = None
         self._beams_already_active = False
-        self._prev_enemies = None
-        self._prev_health = None
         self.was_link_in_cave = False
 
     def reset(self, **kwargs):
@@ -63,8 +61,6 @@ class ZeldaGameWrapper(gym.Wrapper):
         self._last_info = None
         self._location = None
         self._beams_already_active = False
-        self._prev_enemies = None
-        self._prev_health = None
 
         if not self.deterministic:
             for i in range(12):
@@ -95,6 +91,7 @@ class ZeldaGameWrapper(gym.Wrapper):
         ram = unwrapped.get_ram()
         info['buttons'] = self._get_button_names(act, unwrapped.buttons)
         objects = ZeldaObjectData(ram)
+        info['objects'] = objects
 
         link = objects.link
         info['link'] = link
@@ -133,15 +130,7 @@ class ZeldaGameWrapper(gym.Wrapper):
 
         if new_location:
             self._location = location
-            self._prev_health = None
-            self._clear_variables('beam_hits')
-            self._clear_variables('bomb1_hits')
-            self._clear_variables('bomb2_hits')
             self._last_enemies = [None] * 12
-            info['step_hits'] = 0
-        else:
-            # Only check hits if we didn't move room locations
-            info['step_hits'] = self._get_step_hits(act, objects, unwrapped, info)
 
         self._last_info = info
 
@@ -289,50 +278,6 @@ class ZeldaGameWrapper(gym.Wrapper):
     def _get_full_location(self, info):
         return (info['level'], info['location'], is_in_cave(info))
 
-    def _get_step_hits(self, act, objects, unwrapped, info):
-        step_hits = 0
-
-        # capture enemy health data
-        curr_enemy_health = {}
-        for eid in objects.enumerate_enemy_ids():
-            health = objects.get_obj_health(eid)
-
-            # Some enemies, like gels and keese, do not have health.  This makes calculating hits very challenging.
-            # Instead, just set those 0 health enemies to 1 health, which doesn't otherwise affect the game.  The
-            # game will set them to 0 health when they die.
-            if not health and (not self._prev_health or self._prev_health.get(eid, 0) == 0):
-                data = unwrapped.data
-                data.set_value(f'obj_health_{eid:x}', 0x10)
-                health = 1
-
-            curr_enemy_health[eid] = health
-
-        # check if we killed or injured anything
-        if self._prev_health:
-            for eid, health in self._prev_health.items():
-                if eid in curr_enemy_health and curr_enemy_health[eid] < health:
-                    step_hits += 1
-
-        # check if beams, bombs, arrows, etc are active and if they will hit in the future,
-        # as we need to count them as rewards/results of this action so the model trains properly
-        step_hits = self._handle_future_hits(act, info, objects, step_hits, 'beam_hits',
-                                    lambda st: get_beam_state(st) == AnimationState.ACTIVE, self._set_beams_only)
-        step_hits = self._handle_future_hits(act, info, objects, step_hits, 'bomb1_hits',
-                                    lambda st: get_bomb_state(st, 0) == AnimationState.ACTIVE, self._set_bomb1_only)
-        step_hits = self._handle_future_hits(act, info, objects, step_hits, 'bomb2_hits',
-                                    lambda st: get_bomb_state(st, 1) == AnimationState.ACTIVE, self._set_bomb2_only)
-
-        self._prev_health = curr_enemy_health
-        return step_hits
-
-    def _clear_variables(self, name):
-        self._clear_item(name + '_already_active')
-        self._clear_item(name + '_discounted_hits')
-
-    def _clear_item(self, name):
-        if name in self.__dict__:
-            del self.__dict__[name]
-
     def _act_and_wait(self, act):
         action_kind = self._get_action_type(act)
         match action_kind:
@@ -380,6 +325,7 @@ class ZeldaGameWrapper(gym.Wrapper):
         direction = self._get_button_direction(act)
         self._set_direction(direction)
 
+        cooldown = 0
         if action_kind == ActionType.ATTACK:
             obs, rewards, terminated, truncated, info = self.env.step(self._attack_action)
             cooldown = ATTACK_COOLDOWN
@@ -477,92 +423,6 @@ class ZeldaGameWrapper(gym.Wrapper):
             return ActionType.MOVEMENT
 
         return ActionType.NOTHING
-
-    def _handle_future_hits(self, act, info, objects, step_hits, name, condition_check, disable_others):
-        info[name] = 0
-
-        already_active_name = name + '_already_active'
-        discounted_hits = name + '_discounted_hits'
-
-        if condition_check(info):
-            already_active = self.__dict__.get(already_active_name, False)
-            if not already_active:
-                # check if beams will hit something
-                future_hits = self._predict_future(act, info, objects, condition_check, disable_others)
-                info[name] = future_hits
-
-                # count the future hits now, discount them from the later hit
-                step_hits += future_hits
-
-                self.__dict__[discounted_hits] = future_hits
-                self.__dict__[already_active_name] = True
-
-        else:
-            # If we got here, either beams aren't active at all, or we stepped past the end of
-            # the beams.  Make sure we are ready to process them again, and discount any kills
-            # we found.
-            self.__dict__[already_active_name] = False
-
-            # discount hits if we already counted as beam hits
-            discounted_hits = self.__dict__.get(discounted_hits, 0)
-            if discounted_hits and step_hits:
-                discount = min(discounted_hits, step_hits)
-                discounted_hits -= discount
-
-                self.__dict__[discounted_hits] = discounted_hits
-                step_hits -= discount
-
-        return step_hits
-
-    def _predict_future(self, act, info, objects, should_continue, disable_others):
-        # pylint: disable=too-many-locals
-        unwrapped = self.env.unwrapped
-        savestate = unwrapped.em.get_state()
-        data = unwrapped.data
-
-        # disable beams, bombs, or other active damaging effects until the current one is resolved
-        disable_others(data)
-
-        start_enemies = list(objects.enumerate_enemy_ids())
-        start_health = {x: objects.get_obj_health(x) for x in start_enemies}
-
-        # Step over until should_continue is false, or until we left this room or hit a termination condition.
-        # Update info at each iteration.
-        location = (info['level'], info['location'])
-
-        while should_continue(info) and not is_mode_death(info['mode']) and \
-                location == (info['level'], info['location']):
-            data.set_value('hearts_and_containers', 0xff) # make sure we don't die
-
-            _, _, terminated, truncated, info = unwrapped.step(act)
-            if terminated or truncated:
-                break
-
-        hits = 0
-
-        objects = ZeldaObjectData(unwrapped.get_ram())
-        end_health = {x: objects.get_obj_health(x) for x in objects.enumerate_enemy_ids()}
-        for enemy in start_enemies:
-            start = start_health.get(enemy, 0)
-            end = objects.get_obj_health(enemy)
-
-            if enemy not in end_health or end < start:
-                hits += 1
-
-        unwrapped.em.set_state(savestate)
-        return hits
-
-    def _set_beams_only(self, data):
-        data.set_value('bomb_or_flame_animation', 0)
-        data.set_value('bomb_or_flame_animation2', 0)
-
-    def _set_bomb1_only(self, data):
-        data.set_value('beam_animation', 0)
-        data.set_value('bomb_or_flame_animation2', 0)
-
-    def _set_bomb2_only(self, data):
-        data.set_value('beam_animation', 0)
-        data.set_value('bomb_or_flame_animation1', 0)
 
     def _get_button_names(self, act, buttons):
         result = []
